\name{pacificFX-package}
\alias{pacificFX-package}
\alias{pacificFX}
\docType{package}
\title{
Download historical FX data
}
\description{

  Download historical FX data from the PACIFIC Exchange Rate
  Service, run by Werner Antweiler at Sauder School of Business,
  University of British Columbia. Visit the homepage
  \url{http://fx.sauder.ubc.ca/} to learn more.

  \emph{The database is for academic purposes only.} Please see the
  \acronym{FAQ} at \url{http://fx.sauder.ubc.ca/} for the conditions of
  use.

}
\author{
Enrico Schumann

Maintainer: Enrico Schumann <es@enricoschumann.net>
}
\keyword{package}
